chore: add detailed admin view
